lambda/requirements.txt
line-bot-sdk>=3.17.1
openai>=1.0.0
boto3>=1.28.0
pytz>=2023.3
langchain_xai>=0.1.2
